hackathon_backend/pyproject.toml
[project]
name = "fastapi-skeleton"
version = "0.1.0"
description = "Hackathon icin moduler ve olceklenebilir FastAPI iskelet projesi"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "celery-types==0.23.0",
    "celery[redis]>=5.3.1",
    "email-validator>=2.3.0",
    "fastapi",
    "flower>=2.0.1",
    "greenlet>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "pre-commit>=4.5.0",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.44",
    "structlog>=25.5.0",
    "uvicorn[standard]",
    "password-validator>=1.0",
    "httpx>=0.28.1",
    "tenacity>=8.2.3",
]

[dependency-groups]
dev = [
    "bandit>=1.9.2",
    "faker>=38.2.0",
    "hypothesis>=6.148.7",
    "locust>=2.42.6",
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-factoryboy>=2.8.1",
    "pytest-faker>=2.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.8",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
check_untyped_defs = true


[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
# Aktif edilecek kurallar:
# E: Hatalar (Pycodestyle)
# W: Uyarılar
# F: Mantık hataları (Pyflakes)
# I: Import sıralaması (Isort)
# B: Olası bug'lar (Flake8-bugbear)
select = ["E", "W", "F", "I", "B"]
ignore = ["B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
skips = ["B101"] # B101: assert kullanımına izin ver (Testler için gerekli)


[tool.pytest.ini_options]
testpaths = ["tests"]
# Konsol çıktısını detaylandır (-v) ve gereksiz uyarıları gizle
addopts = "-v --disable-warnings"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["app"]
omit = [
    "tests/*",
    "*/__init__.py",
    ".venv/*"
]

[tool.coverage.report]
show_missing = true
# Eğer kapsama oranı %80'in altındaysa hata ver
# fail_under = 80
